backend/requirements.txt
asgiref==3.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-api-key==3.1.0
djangorestframework-bulk==0.2.1
drf-writable-nested==0.7.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.28
packaging==25.0
pillow==10.2.0
protobuf==5.26.0
python-dotenv==1.1.0
pytz==2023.3
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.4.0
